% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_temps.R
\name{add_hourly_temps}
\alias{add_hourly_temps}
\title{Add hourly temps to data frame}
\usage{
add_hourly_temps(data, lat, hours)
}
\arguments{
\item{data}{data frame with columns date, tmin, and tmax}

\item{lat}{location latitude}

\item{hours}{either single value or list of hours of the day to calculate}
}
\value{
data frame with added columns hour and temp
}
\description{
Add hourly temps to data frame
}
\examples{
data.frame(date = seq(as.Date("2000/1/1"), as.Date("2000/1/3"), "days"),
tmin = c(1, 2, 6), tmax = c(10, 12, 15)) |> add_hourly_temps(lat = 45, hours = 0:23)

}
