% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_temps.R
\name{add_sunrise_sunset}
\alias{add_sunrise_sunset}
\title{Add sunrise and sunset times to data frame}
\usage{
add_sunrise_sunset(data, lat)
}
\arguments{
\item{data}{data frame with day of year column named 'doy'}

\item{lat}{location latitude (degrees)}
}
\value{
data frame with added columns sunrise and sunset
}
\description{
Add sunrise and sunset times to data frame
}
\examples{
add_sunrise_sunset(data.frame(doy = 1:366), lat = 45)

}
