% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_wind_speed.R
\name{canopy_wind_speed}
\alias{canopy_wind_speed}
\title{Calculate Windspeed within canopy}
\usage{
canopy_wind_speed(wind_speed, canopy_height = 3, fruit_height = 2, a = 0.4)
}
\arguments{
\item{wind_speed}{Wind speed at 10 m height (m/s)}

\item{canopy_height}{Height of canopy (m), default 3 m}

\item{fruit_height}{Height of fruit  (m), default 2 m}

\item{a}{Constant that determines rates of attenuation within canopy, default is 0.4 for fruit trees}
}
\value{
Wind speed at fruit height in canopy (m/s)
}
\description{
Calculate Windspeed within canopy
}
\examples{
canopy_wind_speed(1)

}
