% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{clear_sky}
\alias{clear_sky}
\title{Clear Sky Radiation from Allen et al.}
\usage{
clear_sky(hour, doy, lat, long, time_zone_long, elevation, freq_h = 1)
}
\arguments{
\item{hour}{hour of day}

\item{doy}{day of year}

\item{lat}{latitude}

\item{long}{longitude}

\item{time_zone_long}{longitude of time zone}

\item{elevation}{elevation (meters)}

\item{freq_h}{frequency of observations in hours defaults to 1 hour (0.5 = half hour)}
}
\value{
Clear sky radiation in W/m2
}
\description{
Clear Sky Radiation from Allen et al.
}
\examples{
clear_sky(0:23, 34, 45, -119, -120, 10)
}
