% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_projections}
\alias{climate_projections}
\title{Climate projections included with MACA}
\format{
A list with 3 climate projections
}
\usage{
climate_projections
}
\description{
A list with 3 climate projections historical, rcp45, rcp85
}
\keyword{datasets}
