% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dew_point.R
\name{dew_point}
\alias{dew_point}
\title{Dew Point Temperature based on air temperature and relative humidity}
\usage{
dew_point(relative_humidity, temp)
}
\arguments{
\item{relative_humidity}{Relative humidity, ranges from zero to one}

\item{temp}{Air temperature (°C)}
}
\value{
Dew point temperature (°C)
}
\description{
Dew Point Temperature based on air temperature and relative humidity
}
\examples{
dew_point(0.75, 35)
}
