% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruit_surface_temp.R
\name{fruit_surface_temp}
\alias{fruit_surface_temp}
\title{Calculate fruit surface temperature}
\usage{
fruit_surface_temp(
  t_air,
  wind_speed,
  s_rad,
  t_dew,
  t_ground = t_air,
  ground_emissivity = 0.97,
  fruit_diameter = 0.08,
  fruit_emissivity = 0.95,
  fruit_reflectance = 0.6,
  fruit_sunlit_prop = 1,
  fruit_groundlit_prop = 0,
  fruit_surface_conductance = 5e-05,
  print_all_vars = FALSE,
  possibly = TRUE
)
}
\arguments{
\item{t_air}{Air temperature (째C)}

\item{wind_speed}{Wind speed (m/s)}

\item{s_rad}{solar radiation (W/m^2)}

\item{t_dew}{Dew point temperature (째C)}

\item{t_ground}{Ground temperature (째C), defaults to air temperature}

\item{ground_emissivity}{Ground emissivity, default 0.97}

\item{fruit_diameter}{Fruit diameter (m), default 0.08 meters}

\item{fruit_emissivity}{Fruit emissivity, default 0.95}

\item{fruit_reflectance}{Fruit reflectance, default 0.6}

\item{fruit_sunlit_prop}{Proportion of fruit area exposed to sunlight,
default 1 for fully exposed fruit}

\item{fruit_groundlit_prop}{Proportion of fruit area exposed to longwave
radiation emitted from ground, default 0 for ground shaded fruit}

\item{fruit_surface_conductance}{Fruit surface conductance to water vapor
diffusion (m/s), default 5E-5 m/s}

\item{print_all_vars}{Prints all energy balance components if TRUE, default FALSE}

\item{possibly}{if TRUE returns NA when energy balance can't be solved
likely due to invalid coefficients, useful for debugging.
Default FALSE}
}
\value{
Fruit surface temperature (째C)
}
\description{
Calculate fruit surface temperature
}
\examples{
fruit_surface_temp(t_air = 33, wind_speed = 1,
s_rad = 800, t_dew = 30)
}
