% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_temps.R
\name{hourly_temp}
\alias{hourly_temp}
\title{Calculate Hourly temperature for any hour of the day}
\usage{
hourly_temp(
  hour,
  tmin,
  tmax,
  tmin_prev,
  tmax_prev,
  tmin_next,
  sunrise,
  sunset,
  sunrise_prev,
  sunset_prev,
  sunrise_next
)
}
\arguments{
\item{hour}{hour of day}

\item{tmin}{minimum temperature of day}

\item{tmax}{maximum temperature of day}

\item{tmin_prev}{minimum temperature of previous day}

\item{tmax_prev}{maximum temperature of previous day}

\item{tmin_next}{minimum temperature of next day}

\item{sunrise}{hour of sunrise}

\item{sunset}{hour of sunset}

\item{sunrise_prev}{hour of sunrise for the previous day}

\item{sunset_prev}{hour of sunset for the previous day}

\item{sunrise_next}{hour of sunrise for the next day}
}
\value{
temperature at hour
}
\description{
Calculate Hourly temperature for any hour of the day
}
\examples{
hourly_temp(0:23, 5, 20, 5, 20, 5, 8, 18, 8, 18, 8)
}
