% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kamaik_maca_path.R
\name{kamiak_maca_path}
\alias{kamiak_maca_path}
\title{Get Kamiak File path to MACA data}
\usage{
kamiak_maca_path(location)
}
\arguments{
\item{location}{file name of location to download in format "data_lat_long"}
}
\value{
base path to file in kamiak
}
\description{
Get Kamiak File path to MACA data
}
\examples{
kamiak_maca_path("data_36.65625_-121.65625")
}
