% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maca_2020_cdl}
\alias{maca_2020_cdl}
\title{CONUS CDL codes for each maca location}
\format{
A data frame with 4218330 rows and 2 variables:
\describe{
\item{location}{file name of maca grid cell with format "data_lat_long"}
\item{cdl_code_2020}{Crop Code from the 2020 cdl contained within each location}
}
}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/Release/index.php}
}
\usage{
maca_2020_cdl
}
\description{
A dataset containing 2020 Cropland Data Layer Codes for all CONUS locations
}
\keyword{datasets}
