% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maca_models}
\alias{maca_models}
\title{A list containing the names of 16 MACA models}
\format{
A list with 16 models
}
\usage{
maca_models
}
\description{
A list containing the names of 16 MACA models
}
\keyword{datasets}
