% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_grid.R
\name{nearest_grid}
\alias{nearest_grid}
\title{return file name of nearest grid point based on 1/16 degree grid}
\usage{
nearest_grid(lat, lon)
}
\arguments{
\item{lat}{input latitude (degrees)}

\item{lon}{input longitude (degrees)}
}
\value{
file name of nearest grid point
}
\description{
return file name of nearest grid point based on 1/16 degree grid
}
\examples{
nearest_grid(46.3758, -123.4567)
}
