% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_binary.R
\name{read_binary}
\alias{read_binary}
\title{Reads a binary weather file}
\usage{
read_binary(file_path, num_vars, hist)
}
\arguments{
\item{file_path}{path to file location}

\item{num_vars}{number of variables in data set either 8 or 4}

\item{hist}{boolean representing time frame of data TRUE for historical
FALSE for future if missing hist will be detected oin file path}
}
\value{
a matrix with date and either 4 or 8 environmental variables
}
\description{
Reads a binary weather file
}
