% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_binary.R
\name{read_binary_addmdy_8var}
\alias{read_binary_addmdy_8var}
\title{Read files with 8 variables}
\usage{
read_binary_addmdy_8var(filename, ymd)
}
\arguments{
\item{filename}{path to file loaction}

\item{ymd}{date column}
}
\value{
binary data
}
\description{
Read files with 8 variables
}
