% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_binary.R
\name{read_gridmet}
\alias{read_gridmet}
\title{Reads a GRIDMET binary weather file}
\usage{
read_gridmet(file_path, begin = 1979, end = 2019, id = NULL)
}
\arguments{
\item{file_path}{path to file location}

\item{begin}{beginning year of data}

\item{end}{ending year of data}

\item{id}{Either a string or 'NULL'. If a string, the output will contain
a variable with that name with the filename(s) as the value.
If 'NULL', the default, no variable will be created.}
}
\value{
a data frame with date and 8 environmental variables
}
\description{
Reads a GRIDMET binary weather file
}
