% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_temps.R
\name{sunrise_sunset}
\alias{sunrise_sunset}
\title{calculate sunrise and sunset}
\usage{
sunrise_sunset(doy, lat)
}
\arguments{
\item{doy}{day of year}

\item{lat}{latitude (degrees)}
}
\value{
data frame with columns sunrise and sunset
}
\description{
calculate sunrise and sunset
}
\examples{
sunrise_sunset(doy = 1:366, lat = 45)

}
