% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_temps.R
\name{temp_after_sunrise}
\alias{temp_after_sunrise}
\title{Calculate temp after Sunrise}
\usage{
temp_after_sunrise(hour, tmin, tmax, sunrise, sunset)
}
\arguments{
\item{hour}{hour of day}

\item{tmin}{minimum temperature of day}

\item{tmax}{maximum temperature of day}

\item{sunrise}{hour of sunrise}

\item{sunset}{hour of sunset}
}
\value{
temperature at hour
}
\description{
Calculate temp after Sunrise
}
\examples{
temp_after_sunrise(hour = 12, tmin = 15, tmax = 33, sunrise = 8.15, sunset = 18.4)
}
