% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_temps.R
\name{temp_after_sunset}
\alias{temp_after_sunset}
\title{Calculate temp after sunset}
\usage{
temp_after_sunset(hour, tmin, tmax, tmin_next, sunrise, sunset, sunrise_next)
}
\arguments{
\item{hour}{hour of day}

\item{tmin}{minimum temperature of day}

\item{tmax}{maximum temperature of day}

\item{tmin_next}{minimum temperature of next day}

\item{sunrise}{hour of sunrise}

\item{sunset}{hour of sunset}

\item{sunrise_next}{hour of sunrise for the next day}
}
\value{
temperature at hour
}
\description{
Calculate temp after sunset
}
\examples{
temp_after_sunset(hour = 12, 15, 33, 17, 8.15, 18.4, 8.2)
}
